vc/requirements.txt
requests==2.31.0 
tqdm==4.65.0        
tenacity==8.2.2       
python-dotenv==1.0.0  
colorama==0.4.6      